% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{CertificationExamStatus}
\alias{CertificationExamStatus}
\title{CertificationExamStatus Object}
\usage{
CertificationExamStatus(type = NULL, numberUsersPass = NULL)
}
\arguments{
\item{type}{The type of certification exam}

\item{numberUsersPass}{The number of people who have passed the certification exam}
}
\value{
CertificationExamStatus object
}
\description{
CertificationExamStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Status for a Google Partners certification exam.
}
