% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{CertificationStatus}
\alias{CertificationStatus}
\title{CertificationStatus Object}
\usage{
CertificationStatus(type = NULL, examStatuses = NULL, isCertified = NULL)
}
\arguments{
\item{type}{The type of the certification}

\item{examStatuses}{List of certification exam statuses}

\item{isCertified}{Whether certification is passing}
}
\value{
CertificationStatus object
}
\description{
CertificationStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Google Partners certification status.
}

