% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_functions.R
\name{clientMessages.log}
\alias{clientMessages.log}
\title{Logs a generic message from the client, such as `Failed to render component`, `Profile page is running slow`, `More than 500 users have accessed this result.`, etc.}
\usage{
clientMessages.log(LogMessageRequest)
}
\arguments{
\item{LogMessageRequest}{The \link{LogMessageRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/partners/}{Google Documentation}

Other LogMessageRequest functions: \code{\link{LogMessageRequest.clientInfo}},
  \code{\link{LogMessageRequest}}
}

