% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_functions.R
\name{companies.get}
\alias{companies.get}
\title{Gets a company.}
\usage{
companies.get(companyId, requestMetadata.userOverrides.ipAddress = NULL,
  requestMetadata.userOverrides.userId = NULL,
  requestMetadata.locale = NULL, requestMetadata.partnersSessionId = NULL,
  requestMetadata.experimentIds = NULL,
  requestMetadata.trafficSource.trafficSourceId = NULL,
  requestMetadata.trafficSource.trafficSubId = NULL, view = NULL,
  orderBy = NULL, currencyCode = NULL, address = NULL)
}
\arguments{
\item{companyId}{The ID of the company to retrieve}

\item{requestMetadata.userOverrides.ipAddress}{IP address to use instead of the user's geo-located IP address}

\item{requestMetadata.userOverrides.userId}{Logged-in user ID to impersonate instead of the user's ID}

\item{requestMetadata.locale}{Locale to use for the current request}

\item{requestMetadata.partnersSessionId}{Google Partners session ID}

\item{requestMetadata.experimentIds}{Experiment IDs the current request belongs to}

\item{requestMetadata.trafficSource.trafficSourceId}{Identifier to indicate where the traffic comes from}

\item{requestMetadata.trafficSource.trafficSubId}{Second level identifier to indicate where the traffic comes from}

\item{view}{The view of `Company` resource to be returned}

\item{orderBy}{How to order addresses within the returned company}

\item{currencyCode}{If the company's budget is in a different currency code than this one, then the converted budget is converted to this currency code}

\item{address}{The address to use for sorting the company's addresses by proximity}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/partners/}{Google Documentation}
}
