% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_functions.R
\name{companies.list}
\alias{companies.list}
\title{Lists companies.}
\usage{
companies.list(requestMetadata.userOverrides.ipAddress = NULL,
  requestMetadata.userOverrides.userId = NULL,
  requestMetadata.locale = NULL, requestMetadata.partnersSessionId = NULL,
  requestMetadata.experimentIds = NULL,
  requestMetadata.trafficSource.trafficSourceId = NULL,
  requestMetadata.trafficSource.trafficSubId = NULL, pageSize = NULL,
  pageToken = NULL, companyName = NULL, view = NULL,
  minMonthlyBudget.currencyCode = NULL, minMonthlyBudget.units = NULL,
  minMonthlyBudget.nanos = NULL, maxMonthlyBudget.currencyCode = NULL,
  maxMonthlyBudget.units = NULL, maxMonthlyBudget.nanos = NULL,
  industries = NULL, services = NULL, languageCodes = NULL,
  address = NULL, orderBy = NULL, gpsMotivations = NULL,
  websiteUrl = NULL)
}
\arguments{
\item{requestMetadata.userOverrides.ipAddress}{IP address to use instead of the user's geo-located IP address}

\item{requestMetadata.userOverrides.userId}{Logged-in user ID to impersonate instead of the user's ID}

\item{requestMetadata.locale}{Locale to use for the current request}

\item{requestMetadata.partnersSessionId}{Google Partners session ID}

\item{requestMetadata.experimentIds}{Experiment IDs the current request belongs to}

\item{requestMetadata.trafficSource.trafficSourceId}{Identifier to indicate where the traffic comes from}

\item{requestMetadata.trafficSource.trafficSubId}{Second level identifier to indicate where the traffic comes from}

\item{pageSize}{Requested page size}

\item{pageToken}{A token identifying a page of results that the server returns}

\item{companyName}{Company name to search for}

\item{view}{The view of the `Company` resource to be returned}

\item{minMonthlyBudget.currencyCode}{The 3-letter currency code defined in ISO 4217}

\item{minMonthlyBudget.units}{The whole units of the amount}

\item{minMonthlyBudget.nanos}{Number of nano (10^-9) units of the amount}

\item{maxMonthlyBudget.currencyCode}{The 3-letter currency code defined in ISO 4217}

\item{maxMonthlyBudget.units}{The whole units of the amount}

\item{maxMonthlyBudget.nanos}{Number of nano (10^-9) units of the amount}

\item{industries}{List of industries the company can help with}

\item{services}{List of services the company can help with}

\item{languageCodes}{List of language codes that company can support}

\item{address}{The address to use when searching for companies}

\item{orderBy}{How to order addresses within the returned companies}

\item{gpsMotivations}{List of reasons for using Google Partner Search to get companies}

\item{websiteUrl}{Website URL that will help to find a better matched company}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/partners/}{Google Documentation}
}
