% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{Company}
\alias{Company}
\title{Company Object}
\usage{
Company(id = NULL, name = NULL, localizedInfos = NULL, locations = NULL,
  convertedMinMonthlyBudget = NULL, originalMinMonthlyBudget = NULL,
  publicProfile = NULL, certificationStatuses = NULL, ranks = NULL,
  websiteUrl = NULL, industries = NULL, services = NULL)
}
\arguments{
\item{id}{The ID of the company}

\item{name}{The name of the company}

\item{localizedInfos}{The list of localized info for the company}

\item{locations}{The list of company locations}

\item{convertedMinMonthlyBudget}{The minimum monthly budget that the company accepts for partner business, converted to the requested currency code}

\item{originalMinMonthlyBudget}{The unconverted minimum monthly budget that the company accepts for partner business}

\item{publicProfile}{Basic information from the company's public profile}

\item{certificationStatuses}{The list of Google Partners certification statuses for the company}

\item{ranks}{Information related to the ranking of the company within the list of companies}

\item{websiteUrl}{URL of the company's website}

\item{industries}{Industries the company can help with}

\item{services}{Services the company can help with}
}
\value{
Company object
}
\description{
Company Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A company resource in the Google Partners API. Once certified, it qualifies for being searched by advertisers.
}
