% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{CreateLeadRequest}
\alias{CreateLeadRequest}
\title{CreateLeadRequest Object}
\usage{
CreateLeadRequest(requestMetadata = NULL, lead = NULL,
  recaptchaChallenge = NULL)
}
\arguments{
\item{requestMetadata}{Current request metadata}

\item{lead}{The lead resource}

\item{recaptchaChallenge}{reCaptcha challenge info}
}
\value{
CreateLeadRequest object
}
\description{
CreateLeadRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for CreateLead.
}

