% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{CreateLeadResponse}
\alias{CreateLeadResponse}
\title{CreateLeadResponse Object}
\usage{
CreateLeadResponse(responseMetadata = NULL, lead = NULL,
  recaptchaStatus = NULL)
}
\arguments{
\item{responseMetadata}{Current response metadata}

\item{lead}{Lead that was created depending on the outcome of reCaptcha validation}

\item{recaptchaStatus}{The outcome of reCaptcha validation}
}
\value{
CreateLeadResponse object
}
\description{
CreateLeadResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for CreateLead. Debug information about this request.
}

