% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{DebugInfo}
\alias{DebugInfo}
\title{DebugInfo Object}
\usage{
DebugInfo(serverInfo = NULL, serviceUrl = NULL, serverTraceInfo = NULL)
}
\arguments{
\item{serverInfo}{Info about the server that serviced this request}

\item{serviceUrl}{URL of the service that handled this request}

\item{serverTraceInfo}{Server-side debug stack trace}
}
\value{
DebugInfo object
}
\description{
DebugInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Debug information about this request.
}

