% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{EventData}
\alias{EventData}
\title{EventData Object}
\usage{
EventData(key = NULL, values = NULL)
}
\arguments{
\item{key}{Data type}

\item{values}{Data values}
}
\value{
EventData object
}
\description{
EventData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Key value data pair for an event.
}
