% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{GetCompanyResponse}
\alias{GetCompanyResponse}
\title{GetCompanyResponse Object}
\usage{
GetCompanyResponse(responseMetadata = NULL, company = NULL)
}
\arguments{
\item{responseMetadata}{Current response metadata}

\item{company}{The company}
}
\value{
GetCompanyResponse object
}
\description{
GetCompanyResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for GetCompany.
}
