% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{Lead}
\alias{Lead}
\title{Lead Object}
\usage{
Lead(id = NULL, type = NULL, email = NULL, givenName = NULL,
  familyName = NULL, websiteUrl = NULL, phoneNumber = NULL,
  comments = NULL, gpsMotivations = NULL, minMonthlyBudget = NULL)
}
\arguments{
\item{id}{ID of the lead}

\item{type}{Type of lead}

\item{email}{Email address of lead source}

\item{givenName}{First name of lead source}

\item{familyName}{Last name of lead source}

\item{websiteUrl}{Website URL of lead source}

\item{phoneNumber}{Phone number of lead source}

\item{comments}{Comments lead source gave}

\item{gpsMotivations}{List of reasons for using Google Partner Search and creating a lead}

\item{minMonthlyBudget}{The minimum monthly budget lead source is willing to spend}
}
\value{
Lead object
}
\description{
Lead Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A lead resource that represents an advertiser contact for a `Company`. These are usually generated via Google Partner Search (the advertiser portal).
}

