% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{ListCompaniesResponse}
\alias{ListCompaniesResponse}
\title{ListCompaniesResponse Object}
\usage{
ListCompaniesResponse(responseMetadata = NULL, companies = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{responseMetadata}{Current response metadata}

\item{companies}{The list of companies}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListCompaniesResponse object
}
\description{
ListCompaniesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ListCompanies.
}
