% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{ListUserStatesResponse}
\alias{ListUserStatesResponse}
\title{ListUserStatesResponse Object}
\usage{
ListUserStatesResponse(responseMetadata = NULL, userStates = NULL)
}
\arguments{
\item{responseMetadata}{Current response metadata}

\item{userStates}{User's states}
}
\value{
ListUserStatesResponse object
}
\description{
ListUserStatesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ListUserStates.
}
