% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{LocalizedCompanyInfo}
\alias{LocalizedCompanyInfo}
\title{LocalizedCompanyInfo Object}
\usage{
LocalizedCompanyInfo(languageCode = NULL, displayName = NULL,
  overview = NULL, countryCodes = NULL)
}
\arguments{
\item{languageCode}{Language code of the localized company info, as defined by BCP 47 (IETF BCP 47, 'Tags for Identifying Languages')}

\item{displayName}{Localized display name}

\item{overview}{Localized brief description that the company uses to advertise themselves}

\item{countryCodes}{List of country codes for the localized company info}
}
\value{
LocalizedCompanyInfo object
}
\description{
LocalizedCompanyInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The localized company information.
}
