% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{Location}
\alias{Location}
\title{Location Object}
\usage{
Location(address = NULL, latLng = NULL)
}
\arguments{
\item{address}{The complete address of the location}

\item{latLng}{The latitude and longitude of the location, in degrees}
}
\value{
Location object
}
\description{
Location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A location with address and geographic coordinates.
}

