% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{LogMessageRequest.clientInfo}
\alias{LogMessageRequest.clientInfo}
\title{LogMessageRequest.clientInfo Object}
\usage{
LogMessageRequest.clientInfo()
}
\value{
LogMessageRequest.clientInfo object
}
\description{
LogMessageRequest.clientInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map of client info, such as URL, browser navigator, browser platform, etc.
}
\seealso{
Other LogMessageRequest functions: \code{\link{LogMessageRequest}},
  \code{\link{clientMessages.log}}
}
