% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{LogMessageRequest}
\alias{LogMessageRequest}
\title{LogMessageRequest Object}
\usage{
LogMessageRequest(LogMessageRequest.clientInfo = NULL,
  requestMetadata = NULL, level = NULL, details = NULL,
  clientInfo = NULL)
}
\arguments{
\item{LogMessageRequest.clientInfo}{The \link{LogMessageRequest.clientInfo} object or list of objects}

\item{requestMetadata}{Current request metadata}

\item{level}{Message level of client message}

\item{details}{Details about the client message}

\item{clientInfo}{Map of client info, such as URL, browser navigator, browser platform, etc}
}
\value{
LogMessageRequest object
}
\description{
LogMessageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for LogClientMessage.
}
\seealso{
Other LogMessageRequest functions: \code{\link{LogMessageRequest.clientInfo}},
  \code{\link{clientMessages.log}}
}
