% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{LogMessageResponse}
\alias{LogMessageResponse}
\title{LogMessageResponse Object}
\usage{
LogMessageResponse(responseMetadata = NULL)
}
\arguments{
\item{responseMetadata}{Current response metadata}
}
\value{
LogMessageResponse object
}
\description{
LogMessageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for LogClientMessage.
}
