% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{LogUserEventRequest}
\alias{LogUserEventRequest}
\title{Google Partners API Objects 
Lets advertisers search certified companies and create contact leads with them, and also audits the usage of clients.}
\usage{
LogUserEventRequest(requestMetadata = NULL, eventAction = NULL,
  eventCategory = NULL, eventScope = NULL, eventDatas = NULL,
  url = NULL, lead = NULL)
}
\arguments{
\item{requestMetadata}{Current request metadata}

\item{eventAction}{The action that occurred}

\item{eventCategory}{The category the action belongs to}

\item{eventScope}{The scope of the event}

\item{eventDatas}{List of event data for the event}

\item{url}{The URL where the event occurred}

\item{lead}{Advertiser lead information}
}
\value{
LogUserEventRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:57:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlepartnersv2.auto/R/partners_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
LogUserEventRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for LogUserEvent.
}
\seealso{
Other LogUserEventRequest functions: \code{\link{userEvents.log}}
}
