% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{LogUserEventResponse}
\alias{LogUserEventResponse}
\title{LogUserEventResponse Object}
\usage{
LogUserEventResponse(responseMetadata = NULL)
}
\arguments{
\item{responseMetadata}{Current response metadata}
}
\value{
LogUserEventResponse object
}
\description{
LogUserEventResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for LogUserEvent.
}
