% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{Money}
\alias{Money}
\title{Money Object}
\usage{
Money(currencyCode = NULL, units = NULL, nanos = NULL)
}
\arguments{
\item{currencyCode}{The 3-letter currency code defined in ISO 4217}

\item{units}{The whole units of the amount}

\item{nanos}{Number of nano (10^-9) units of the amount}
}
\value{
Money object
}
\description{
Money Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an amount of money with its currency type.
}
