% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_functions.R
\docType{package}
\name{partners_googleAuthR}
\alias{partners_googleAuthR}
\alias{partners_googleAuthR-package}
\title{Google Partners API
Lets advertisers search certified companies and create contact leads with them, and also audits the usage of clients.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:21:00
filename: /Users/mark/dev/R/autoGoogleAPI/googlepartnersv2.auto/R/partners_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

