% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{PublicProfile}
\alias{PublicProfile}
\title{PublicProfile Object}
\usage{
PublicProfile(id = NULL, displayName = NULL, url = NULL,
  displayImageUrl = NULL)
}
\arguments{
\item{id}{The ID which can be used to retrieve more details about the public profile}

\item{displayName}{The display name of the public profile}

\item{url}{The URL of the public profile}

\item{displayImageUrl}{The URL to the main display image of the public profile}
}
\value{
PublicProfile object
}
\description{
PublicProfile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic information from a public profile.
}

