% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{Rank}
\alias{Rank}
\title{Rank Object}
\usage{
Rank(type = NULL, value = NULL)
}
\arguments{
\item{type}{The type of rank}

\item{value}{The numerical value of the rank}
}
\value{
Rank object
}
\description{
Rank Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information related to ranking of results.
}
