% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{RecaptchaChallenge}
\alias{RecaptchaChallenge}
\title{RecaptchaChallenge Object}
\usage{
RecaptchaChallenge(id = NULL, response = NULL)
}
\arguments{
\item{id}{The ID of the reCaptcha challenge}

\item{response}{The response to the reCaptcha challenge}
}
\value{
RecaptchaChallenge object
}
\description{
RecaptchaChallenge Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
reCaptcha challenge info.
}

