% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{RequestMetadata}
\alias{RequestMetadata}
\title{RequestMetadata Object}
\usage{
RequestMetadata(userOverrides = NULL, locale = NULL,
  partnersSessionId = NULL, experimentIds = NULL, trafficSource = NULL)
}
\arguments{
\item{userOverrides}{Values to use instead of the user's respective defaults for the current request}

\item{locale}{Locale to use for the current request}

\item{partnersSessionId}{Google Partners session ID}

\item{experimentIds}{Experiment IDs the current request belongs to}

\item{trafficSource}{Source of traffic for the current request}
}
\value{
RequestMetadata object
}
\description{
RequestMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common data that is in each API request.
}

