% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{ResponseMetadata}
\alias{ResponseMetadata}
\title{ResponseMetadata Object}
\usage{
ResponseMetadata(debugInfo = NULL)
}
\arguments{
\item{debugInfo}{Debug information about this request}
}
\value{
ResponseMetadata object
}
\description{
ResponseMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common data that is in each API response.
}
