% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{TrafficSource}
\alias{TrafficSource}
\title{TrafficSource Object}
\usage{
TrafficSource(trafficSourceId = NULL, trafficSubId = NULL)
}
\arguments{
\item{trafficSourceId}{Identifier to indicate where the traffic comes from}

\item{trafficSubId}{Second level identifier to indicate where the traffic comes from}
}
\value{
TrafficSource object
}
\description{
TrafficSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Source of traffic for the current request.
}
