% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_functions.R
\name{userEvents.log}
\alias{userEvents.log}
\title{Logs a user event.}
\usage{
userEvents.log(LogUserEventRequest)
}
\arguments{
\item{LogUserEventRequest}{The \link{LogUserEventRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/partners/}{Google Documentation}

Other LogUserEventRequest functions: \code{\link{LogUserEventRequest}}
}

