% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners_objects.R
\name{UserOverrides}
\alias{UserOverrides}
\title{UserOverrides Object}
\usage{
UserOverrides(ipAddress = NULL, userId = NULL)
}
\arguments{
\item{ipAddress}{IP address to use instead of the user's geo-located IP address}

\item{userId}{Logged-in user ID to impersonate instead of the user's ID}
}
\value{
UserOverrides object
}
\description{
UserOverrides Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Values to use instead of the user's respective defaults. These are only honored by whitelisted products.
}
